st-hal: Fix ubsan mul-overflow in buffer_thread_loop
- convert_bytes_to_ms: input_bytes * 1000 stays as 32-bit unsigned int.
- With bytes_written accumulating during buffering, this overflows UINT32_MAX (~4.3
  MB, i.e., ~134 seconds at 16kHz/16bit/mono, much sooner at 48kHz/stereo).
- convert_ms_to_bytes: input_ms * rate * channels * bps chain is all 32-bit.
  At 48kHz/stereo/32-bit PCM the multiplier is 384, overflowing at ~11 seconds
  worth of input_ms.
- fix: cast the first operand to uint64_t in both helpers so the entire
  multiplication chain uses 64-bit arithmetic, then truncate back to unsigned int
  at the return.
- second_stage:cast operands to uint32_t explicitly with bounds checking, or use
  int32_t throughout since kw_start_idx is uint32_t and negative positions
  should be clamped to 0.

+++ st_hw_common.h
@@ -29,6 +29,10 @@
  * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
  * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN
  * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ * Changes from Qualcomm Technologies, Inc. are provided under the following license:
+ * Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+ * SPDX-License-Identifier: BSD-3-Clause-Clear
  */
 
 #include "sound_trigger_hw.h"
@@ -67,7 +71,7 @@ static inline uint64_t get_current_time_ns()
     struct timespec ts;
 
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return (ts.tv_sec * NSECS_PER_SEC) + ts.tv_nsec;
+    return ((uint64_t)ts.tv_sec * (uint64_t)NSECS_PER_SEC) + (uint64_t)ts.tv_nsec;
 }
 
 static inline unsigned int convert_ms_to_bytes
@@ -76,7 +80,7 @@ static inline unsigned int convert_ms_to_bytes
     struct pcm_config *config
 )
 {
-    return ((input_ms * config->rate * config->channels *
+    return (unsigned int)(((uint64_t)input_ms * config->rate * config->channels *
         (pcm_format_to_bits(config->format) >> 3)) / 1000);
 }
 
@@ -86,6 +90,7 @@ static inline unsigned int convert_bytes_to_ms
     struct pcm_config *config
 )
 {
-    return ((input_bytes * 1000) / (config->rate * config->channels *
+    return (unsigned int)(((uint64_t)input_bytes * 1000) /
+        ((uint64_t)config->rate * config->channels *
         (pcm_format_to_bits(config->format) >> 3)));
 }
